clean eunomia-rs and move common rs lib to eunomia-rs
