+++ compiler/cmd/Cargo.toml
@@ -10,6 +10,7 @@ clap = { git = "https://github.com/clap-rs/clap.git", rev = "3e7ca92", features 
 regex = "1"
 run_script = "^0.10.0"
 anyhow = "1.0"
+eunomia-rs = {path = "../../eunomia-sdks/eunomia-rs"}
 home = "0.5.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
