+++ eunomia-sdks/eunomia-rs/Cargo.toml
@@ -4,13 +4,10 @@ version = "0.1.0"
 edition = "2021"
 description = "A rust binding for eunomia-bpf."
 license = "MIT"
-build = "build.rs"
 
 [dependencies]
-link-cplusplus = "1.0"
-
-[build-dependencies]
-bindgen = "0.60.1"
+anyhow = "1.0"
+home = "0.5.4"
 
 [lib]
 crate-type = ["lib", "staticlib"]
