--- eunomia-sdks/eunomia-rs/wrapper.h
@@ -1,1 +1,0 @@
-../../eunomia-bpf/include/eunomia/eunomia-bpf.h
